Tighten dependency specifications for ex_doc and earmark.
Previous spec for ex_doc allowed upgrade to 0.14.5, which would be fine, but caused a failure in `mix deps.get` because specified version of earmark is too old.

But ... we can't upgrade to current earmark (1.1.1 as of this writing) because it deprecates a function that latest ex_doc uses. So we specify latest compatible version.